+++ CotEditor/Sources/String+Counting.swift
@@ -9,7 +9,7 @@
  
  ------------------------------------------------------------------------------
  
- © 2014-2016 1024jp
+ © 2014-2017 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -32,14 +32,22 @@ extension String {
     /// number of composed characters in the whole string
     var numberOfComposedCharacters: Int {
         
+        return self.countComposedCharacters()
+    }
+    
+    
+    ///
+    func countComposedCharacters(_ body: ((_ stop: inout Bool) -> (Void))? = nil) -> Int {
+        
         guard !self.isEmpty else { return 0 }
         
         // normalize using NFC
         let string = self.precomposedStringWithCanonicalMapping
         
         var count = 0
-        string.enumerateSubstrings(in: string.startIndex..<string.endIndex, options: [.byComposedCharacterSequences, .substringNotRequired]) { (substring, substringRange, enclosingRange, stop) in
+        string.enumerateSubstrings(in: string.startIndex..<string.endIndex, options: [.byComposedCharacterSequences, .substringNotRequired]) { (_, _, _, stop) in
             count += 1
+            body?(&stop)
         }
         
         return count